Finalisation de l'application

+++ app/src/main/java/fr/diginamic/formation/superquizz/database/QuestionsDatabaseHelper.java
@@ -2,9 +2,11 @@ package fr.diginamic.formation.superquizz.database;
 
 import android.content.ContentValues;
 import android.content.Context;
+import android.content.SharedPreferences;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
+import android.preference.PreferenceManager;
 import android.util.Log;
 
 import java.io.IOException;
@@ -12,17 +14,14 @@ import java.util.ArrayList;
 
 import fr.diginamic.formation.superquizz.model.Question;
 import fr.diginamic.formation.superquizz.model.TypeQuestion;
+import fr.diginamic.formation.superquizz.ui.activities.PreferencesActivity;
 
 public class QuestionsDatabaseHelper extends SQLiteOpenHelper implements APIClient.APIResult<ArrayList<Question>> {
     private static QuestionsDatabaseHelper questionsDatabaseHelper;
-    // Database Info
+    private static SharedPreferences mSettings;
     private static final String DATABASE_NAME = "questionsDatabase";
     private static final int DATABASE_VERSION = 1;
-
-    // Table Names
     private static final String TABLE_QCM = "qcm";
-
-    // Post Table Columns
     private static final String KEY_QUESTION_ID = "question_id";
     private static final String ENTITLE = "question";
     private static final String ANSWER_1 = "answer_1";
@@ -36,6 +35,7 @@ public class QuestionsDatabaseHelper extends SQLiteOpenHelper implements APIClie
     public static synchronized QuestionsDatabaseHelper getInstance(Context context) {
         if (questionsDatabaseHelper == null) {
             questionsDatabaseHelper = new QuestionsDatabaseHelper(context.getApplicationContext());
+            mSettings = PreferenceManager.getDefaultSharedPreferences(context);
         }
         return questionsDatabaseHelper;
     }
@@ -90,8 +90,11 @@ public class QuestionsDatabaseHelper extends SQLiteOpenHelper implements APIClie
             db.insertOrThrow(TABLE_QCM, null, values);
             db.setTransactionSuccessful();
 
-            if(addOnline){
-                APIClient.getInstance().addQuestion(question);
+            boolean onlinePreference = mSettings.getBoolean(PreferencesActivity.SAVE_ONLINE, true);
+            if(onlinePreference){
+                if(addOnline){
+                    APIClient.getInstance().addQuestion(question);
+                }
             }
         } catch (Exception e) {
             Log.e("DataBase ERROR", "Error while trying to add question to database");
@@ -132,56 +135,6 @@ public class QuestionsDatabaseHelper extends SQLiteOpenHelper implements APIClie
         return questions;
     }
 
-    public ArrayList<Question> getAllQuestionsNotAnswered() {
-        ArrayList<Question> questions = new ArrayList<>();
-
-        String QCM_SELECT_QUERY = String.format("SELECT * FROM %s WHERE %s IS NULL", TABLE_QCM, USER_ANSWER);
-
-        SQLiteDatabase db = getReadableDatabase();
-        Cursor cursor = db.rawQuery(QCM_SELECT_QUERY, null);
-        try {
-            if (cursor.moveToFirst()) {
-                do {
-                    Question newQuestion = new Question(cursor.getString(cursor.getColumnIndex(ENTITLE)));
-                    newQuestion.addProposition(cursor.getString(cursor.getColumnIndex(ANSWER_1)));
-                    newQuestion.addProposition(cursor.getString(cursor.getColumnIndex(ANSWER_2)));
-                    newQuestion.addProposition(cursor.getString(cursor.getColumnIndex(ANSWER_3)));
-                    newQuestion.addProposition(cursor.getString(cursor.getColumnIndex(ANSWER_4)));
-                    newQuestion.setGoodAnswer(cursor.getString(cursor.getColumnIndex(GOOD_ANSWER)));
-                    newQuestion.setId(cursor.getInt(cursor.getColumnIndex(KEY_QUESTION_ID)));
-                    newQuestion.setType(TypeQuestion.SIMPLE);
-
-                    questions.add(newQuestion);
-                } while(cursor.moveToNext());
-            }
-        } catch (Exception e) {
-            Log.e("DataBase ERROR", "Error while trying to get questions from database");
-        } finally {
-            if (cursor != null && !cursor.isClosed()) {
-                cursor.close();
-            }
-        }
-        return questions;
-    }
-
-    public int updateUserAnswer(Question question, String userAnswer,  int id) {
-        SQLiteDatabase db = this.getWritableDatabase();
-
-        ContentValues values = new ContentValues();
-        values.put(ENTITLE, question.getEntitle());
-        values.put(ANSWER_1, question.getProposition(0));
-        values.put(ANSWER_2, question.getProposition(1));
-        values.put(ANSWER_3, question.getProposition(2));
-        values.put(ANSWER_4, question.getProposition(3));
-        values.put(USER_ANSWER, userAnswer);
-        values.put(GOOD_ANSWER, question.getGoodAnswer());
-
-        APIClient.getInstance().updateQuestion(question, id);
-
-        return db.update(TABLE_QCM, values, KEY_QUESTION_ID + " = ?",
-                new String[] { String.valueOf(id)});
-    }
-
     public int updateQuestion(Question question, int id) {
         SQLiteDatabase db = this.getWritableDatabase();
 
@@ -193,7 +146,10 @@ public class QuestionsDatabaseHelper extends SQLiteOpenHelper implements APIClie
         values.put(ANSWER_4, question.getProposition(3));
         values.put(GOOD_ANSWER, question.getGoodAnswer());
 
-        APIClient.getInstance().updateQuestion(question, id);
+        boolean onlinePreference = mSettings.getBoolean(PreferencesActivity.SAVE_ONLINE, true);
+        if(onlinePreference){
+            APIClient.getInstance().updateQuestion(question, id);
+        }
 
         return db.update(TABLE_QCM, values, KEY_QUESTION_ID + " = ?",
                 new String[] { String.valueOf(id)});
@@ -208,7 +164,10 @@ public class QuestionsDatabaseHelper extends SQLiteOpenHelper implements APIClie
         try {
             db.delete(TABLE_QCM, clause, clauseArgs);
             db.setTransactionSuccessful();
-            APIClient.getInstance().deleteQuestion(question);
+            boolean onlinePreference = mSettings.getBoolean(PreferencesActivity.SAVE_ONLINE, true);
+            if(onlinePreference){
+                APIClient.getInstance().deleteQuestion(question);
+            }
         } catch (Exception e) {
             Log.d("DataBase ERROR", "Error while trying to delete all questions");
         } finally {
